_tools/requirements.txt
# Auto-maintained by Ayra. Tools dependencies.
pytest>=7.0
pyyaml>=6.0
